requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
scikit-learn
numpy
openai
requests
beautifulsoup4
selenium
python-dotenv
